+++ Skills/MagicDamage.cs
@@ -25,7 +25,7 @@ namespace Infinitum.Skills
             Cost = 60;
             EffectBuff = 0;
             MultiplierEffect = 0.01f;
-            Type = (int)SkillEnums.Type.PostUpdateEquips;
+            Type = SkillEnums.Type.PostUpdateEquips;
         }
     }
 }
