+++ Skills/SummonDamage.cs
@@ -20,7 +20,7 @@ namespace Infinitum.Skills
             Cost = 60;
             EffectBuff = 0;
             MultiplierEffect = 0.01f;
-            Type = (int)SkillEnums.Type.PostUpdateEquips;
+            Type = SkillEnums.Type.PostUpdateEquips;
         }
         public override void ApplyStatToPlayer()
         {
